requirements.txt
pandas
numpy
python-dotenv
dotenv 
openpyxl
pyyaml
customtkinter
xlrd
chardet
charset_normalizer
pillow
pyinstaller
pytest
yagmail==0.15.293
jinja2==3.1.2
reportlab==4.0.4